refactor: split type checking
