+++ src/semantics/check-types/check-binaryen-call.ts
@@ -1,0 +1,6 @@
+import { Call } from "../../syntax-objects/call.js";
+
+export const checkBinaryenCall = (call: Call) => {
+  return call; // TODO: Actually check?
+};
+
