+++ src/semantics/check-types/check-fn.ts
@@ -1,0 +1,44 @@
+import { Fn } from "../../syntax-objects/fn.js";
+import { typesAreCompatible } from "../resolution/index.js";
+import { checkParameters } from "./check-parameters.js";
+import { checkTypes } from "./check-types.js";
+import { checkTypeExpr } from "./check-type-expr.js";
+
+export const checkFnTypes = (fn: Fn): Fn => {
+  if (fn.genericInstances) {
+    fn.genericInstances.forEach(checkFnTypes);
+    return fn;
+  }
+
+  // If the function has type parameters and not genericInstances, it isn't in use and wont be compiled.
+  if (fn.typeParameters) {
+    return fn;
+  }
+
+  checkParameters(fn.parameters);
+  checkTypes(fn.body);
+
+  if (fn.returnTypeExpr) {
+    checkTypeExpr(fn.returnTypeExpr);
+  }
+
+  if (!fn.returnType) {
+    throw new Error(
+      `Unable to determine return type for ${fn.name} at ${fn.location}`
+    );
+  }
+
+  const inferredReturnType = fn.inferredReturnType;
+
+  if (
+    inferredReturnType &&
+    !typesAreCompatible(inferredReturnType, fn.returnType)
+  ) {
+    throw new Error(
+      `Fn, ${fn.name}, return value type (${inferredReturnType?.name}) is not compatible with annotated return type (${fn.returnType?.name}) at ${fn.location}`
+    );
+  }
+
+  return fn;
+};
+
