+++ src/semantics/check-types/check-union-type.ts
@@ -1,0 +1,25 @@
+import { UnionType, Type } from "../../syntax-objects/types.js";
+import { checkTypeExpr } from "./check-type-expr.js";
+
+export const checkUnionType = (union: UnionType) => {
+  union.childTypeExprs.each(checkTypeExpr);
+
+  if (union.types.length !== union.childTypeExprs.length) {
+    throw new Error(`Unable to resolve every type in union ${union.location}`);
+  }
+
+  union.types.forEach((t: Type) => {
+    const isObjectType =
+      t.isObjectType() ||
+      t.isIntersectionType() ||
+      t.isUnionType();
+    if (!isObjectType) {
+      throw new Error(
+        `Union must be made up of object types ${union.location}`
+      );
+    }
+  });
+
+  return union;
+};
+
