+++ src/semantics/check-types/check-use.ts
@@ -1,0 +1,7 @@
+import { Use } from "../../syntax-objects/use.js";
+
+export const checkUse = (use: Use) => {
+  // TODO: Maybe check for dupes or some
+  return use;
+};
+
