+++ src/semantics/check-types/index.ts
@@ -1,0 +1,6 @@
+export { checkTypes } from "./check-types.js";
+export { checkAssign } from "./check-assign.js";
+export { checkIf } from "./check-if.js";
+export { checkBinaryenCall } from "./check-binaryen-call.js";
+export { checkLabeledArg } from "./check-labeled-arg.js";
+export { assertValidExtension } from "./check-object-type.js";
